Refactored code, BaseBot class added
