+++ jbot/src/main/java/me/ramswaroop/jbot/core/slack/Bot.java
@@ -1,6 +1,7 @@
 package me.ramswaroop.jbot.core.slack;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import me.ramswaroop.jbot.core.common.BaseBot;
 import me.ramswaroop.jbot.core.common.BotWebSocketHandler;
 import me.ramswaroop.jbot.core.common.Controller;
 import me.ramswaroop.jbot.core.common.EventType;
@@ -9,7 +10,6 @@ import me.ramswaroop.jbot.core.slack.models.Message;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.util.StringUtils;
 import org.springframework.web.socket.CloseStatus;
 import org.springframework.web.socket.TextMessage;
 import org.springframework.web.socket.WebSocketHandler;
@@ -19,10 +19,6 @@ import org.springframework.web.socket.client.standard.StandardWebSocketClient;
 
 import javax.annotation.PostConstruct;
 import java.io.IOException;
-import java.lang.reflect.Method;
-import java.util.*;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
 /**
  * Base class for making Slack Bots. Any class extending
@@ -31,29 +27,9 @@ import java.util.regex.Pattern;
  * @author ramswaroop
  * @version 1.0.0, 05/06/2016
  */
-public abstract class Bot {
+public abstract class Bot extends BaseBot {
 
     private static final Logger logger = LoggerFactory.getLogger(Bot.class);
-    /**
-     * A Map of all methods annotated with {@link Controller} where key is the {@link EventType#name()} and
-     * value is a list of {@link MethodWrapper}. NOTE: It does not contain methods which are part of any
-     * conversation.
-     */
-    private final Map<String, List<MethodWrapper>> eventToMethodsMap = new HashMap<>();
-    /**
-     * A Map of all methods annotated with {@link Controller} where key is the {@link Method#getName()} and
-     * value is the respective {@link MethodWrapper}.
-     */
-    private final Map<String, MethodWrapper> methodNameMap = new HashMap<>();
-    /**
-     * A List of names of the methods which are part of any conversation.
-     */
-    private final List<String> conversationMethodNames = new ArrayList<>();
-    /**
-     * A List of Queues with each Queue holding all methods for a particular conversation. Methods
-     * can be chained into a conversation by {@link Controller#next()}.
-     */
-    private final Map<String, Queue<MethodWrapper>> conversationQueueMap = new HashMap<>();
 
     /**
      * Service to access Slack APIs.
@@ -76,44 +52,6 @@ public abstract class Bot {
      * @return
      */
     public abstract Bot getSlackBot();
-
-    /**
-     * Construct a map of all the controller methods to handle RTM Events.
-     */
-    public Bot() {
-        Method[] methods = this.getClass().getMethods();
-        for (Method method : methods) {
-            if (method.isAnnotationPresent(Controller.class)) {
-                Controller controller = method.getAnnotation(Controller.class);
-                EventType[] eventTypes = controller.events();
-                String pattern = controller.pattern();
-                String next = controller.next();
-
-                if (!StringUtils.isEmpty(next)) {
-                    conversationMethodNames.add(next);
-                }
-
-                MethodWrapper methodWrapper = new MethodWrapper();
-                methodWrapper.setMethod(method);
-                methodWrapper.setPattern(pattern);
-                methodWrapper.setNext(next);
-
-                if (!conversationMethodNames.contains(method.getName())) {
-                    for (EventType eventType : eventTypes) {
-                        List<MethodWrapper> methodWrappers = eventToMethodsMap.get(eventType.name());
-
-                        if (methodWrappers == null) {
-                            methodWrappers = new ArrayList<>();
-                        }
-
-                        methodWrappers.add(methodWrapper);
-                        eventToMethodsMap.put(eventType.name(), methodWrappers);
-                    }
-                }
-                methodNameMap.put(method.getName(), methodWrapper);
-            }
-        }
-    }
 
     /**
      * Invoked after a successful web socket connection is
@@ -185,49 +123,6 @@ public abstract class Bot {
     }
 
     /**
-     * Call this method to start a conversation.
-     *
-     * @param event
-     */
-    public void startConversation(Event event, String methodName) {
-        String channelId = event.getChannelId();
-
-        if (!StringUtils.isEmpty(channelId)) {
-            Queue<MethodWrapper> queue = formConversationQueue(new LinkedList<>(), methodName);
-            conversationQueueMap.put(channelId, queue);
-        }
-    }
-
-    /**
-     * Call this method to jump to the next method in a conversation.
-     *
-     * @param event
-     */
-    public void nextConversation(Event event) {
-        Queue<MethodWrapper> queue = conversationQueueMap.get(event.getChannelId());
-        if (queue != null) queue.poll();
-    }
-
-    /**
-     * Call this method to stop the end the conversation.
-     *
-     * @param event
-     */
-    public void stopConversation(Event event) {
-        conversationQueueMap.remove(event.getChannelId());
-    }
-
-    /**
-     * Check whether a conversation is up in a particular slack channel.
-     *
-     * @param event
-     * @return true if a conversation is on, false otherwise.
-     */
-    public boolean isConversationOn(Event event) {
-        return conversationQueueMap.get(event.getChannelId()) != null;
-    }
-
-    /**
      * Method to send a reply back to Slack after receiving an {@link Event}.
      * Learn <a href="https://api.slack.com/rtm">more on sending responses to Slack.</a>
      *
@@ -252,31 +147,40 @@ public abstract class Bot {
     }
 
     /**
-     * Encode the text before sending to Slack.
-     * Learn <a href="https://api.slack.com/docs/formatting">more on message formatting in Slack</a>
+     * Call this method to start a conversation.
      *
-     * @param message
-     * @return encoded text.
+     * @param event
      */
-    private String encode(String message) {
-        return message == null ? null : message.replace("&", "&amp;").replace("<", "&lt;").replace(">", "&gt;");
+    public final void startConversation(Event event, String methodName) {
+        startConversation(event.getChannelId(), methodName);
     }
 
     /**
-     * Form a Queue with all the methods responsible for a particular conversation.
+     * Call this method to jump to the next method in a conversation.
      *
-     * @param queue
-     * @param methodName
-     * @return
+     * @param event
      */
-    private Queue<MethodWrapper> formConversationQueue(Queue<MethodWrapper> queue, String methodName) {
-        MethodWrapper methodWrapper = methodNameMap.get(methodName);
-        queue.add(methodWrapper);
-        if (StringUtils.isEmpty(methodName)) {
-            return queue;
-        } else {
-            return formConversationQueue(queue, methodWrapper.getNext());
-        }
+    public final void nextConversation(Event event) {
+        nextConversation(event.getChannelId());
+    }
+
+    /**
+     * Call this method to stop the end the conversation.
+     *
+     * @param event
+     */
+    public final void stopConversation(Event event) {
+        stopConversation(event.getChannelId());
+    }
+
+    /**
+     * Check whether a conversation is up in a particular slack channel.
+     *
+     * @param event
+     * @return true if a conversation is on, false otherwise.
+     */
+    public final boolean isConversationOn(Event event) {
+        return isConversationOn(event.getChannelId());
     }
 
     /**
@@ -287,30 +191,7 @@ public abstract class Bot {
      * @param event
      */
     private void invokeMethods(WebSocketSession session, Event event) {
-        try {
-            List<MethodWrapper> methodWrappers = eventToMethodsMap.get(event.getType().toUpperCase());
-            if (methodWrappers == null) return;
-
-            methodWrappers = new ArrayList<>(methodWrappers);
-            MethodWrapper matchedMethod = getMethodWithMatchingPatternAndFilterUnmatchedMethods(event, methodWrappers);
-            if (matchedMethod != null) {
-                methodWrappers = new ArrayList<>();
-                methodWrappers.add(matchedMethod);
-            }
-
-            if (methodWrappers != null) {
-                for (MethodWrapper methodWrapper : methodWrappers) {
-                    Method method = methodWrapper.getMethod();
-                    if (method.getParameterCount() == 3) {
-                        method.invoke(this, session, event, methodWrapper.getMatcher());
-                    } else {
-                        method.invoke(this, session, event);
-                    }
-                }
-            }
-        } catch (Exception e) {
-            logger.error("Error invoking controller: ", e);
-        }
+        invokeMethods(event, new Object[]{session, event}); // order of arguments must match that of controller method
     }
 
     /**
@@ -320,57 +201,19 @@ public abstract class Bot {
      * @param event
      */
     private void invokeChainedMethod(WebSocketSession session, Event event) {
-        Queue<MethodWrapper> queue = conversationQueueMap.get(event.getChannelId());
-
-        if (queue != null && !queue.isEmpty()) {
-            MethodWrapper methodWrapper = queue.peek();
-
-            try {
-                EventType[] eventTypes = methodWrapper.getMethod().getAnnotation(Controller.class).events();
-                for (EventType eventType : eventTypes) {
-                    if (eventType.name().equals(event.getType().toUpperCase())) {
-                        methodWrapper.getMethod().invoke(this, session, event);
-                        return;
-                    }
-                }
-            } catch (Exception e) {
-                logger.error("Error invoking chained method: ", e);
-            }
-        }
+        invokeChainedMethod(event.getChannelId(), event, new Object[]{session, event}); // order of arguments must 
+                                                                                        // match that of controller method
     }
 
     /**
-     * Search for a method whose {@link Controller#pattern()} match with the {@link Event#text}
-     * in events received from Slack and also filter out the methods whose {@link Controller#pattern()} do not
-     * match with slack message received ({@link Event#text}) for cases where there are no matched methods.
+     * Encode the text before sending to Slack.
+     * Learn <a href="https://api.slack.com/docs/formatting">more on message formatting in Slack</a>
      *
-     * @param event
-     * @param methodWrappers
-     * @return the MethodWrapper whose method pattern match with that of the slack message received, {@code null} if no
-     * such method is found.
+     * @param message
+     * @return encoded text.
      */
-    private MethodWrapper getMethodWithMatchingPatternAndFilterUnmatchedMethods(Event event, List<MethodWrapper> methodWrappers) {
-        if (methodWrappers != null) {
-            Iterator<MethodWrapper> listIterator = methodWrappers.listIterator();
-
-            while (listIterator.hasNext()) {
-                MethodWrapper methodWrapper = listIterator.next();
-                String pattern = methodWrapper.getPattern();
-                String text = event.getText();
-
-                if (!StringUtils.isEmpty(pattern) && !StringUtils.isEmpty(text)) {
-                    Pattern p = Pattern.compile(pattern);
-                    Matcher m = p.matcher(text);
-                    if (m.find()) {
-                        methodWrapper.setMatcher(m);
-                        return methodWrapper;
-                    } else {
-                        listIterator.remove();  // remove methods from the original list whose pattern do not match
-                    }
-                }
-            }
-        }
-        return null;
+    private String encode(String message) {
+        return message == null ? null : message.replace("&", "&amp;").replace("<", "&lt;").replace(">", "&gt;");
     }
 
     private StandardWebSocketClient client() {
@@ -393,71 +236,6 @@ public abstract class Bot {
             manager.start();
         } else {
             logger.error("No websocket url returned by Slack.");
-        }
-    }
-
-    /**
-     * Wrapper class for methods annotated with {@link Controller}.
-     */
-    private class MethodWrapper {
-        private Method method;
-        private String pattern;
-        private Matcher matcher;
-        private String next;
-
-        public Method getMethod() {
-            return method;
-        }
-
-        public void setMethod(Method method) {
-            this.method = method;
-        }
-
-        public String getPattern() {
-            return pattern;
-        }
-
-        public void setPattern(String pattern) {
-            this.pattern = pattern;
-        }
-
-        public Matcher getMatcher() {
-            return matcher;
-        }
-
-        public void setMatcher(Matcher matcher) {
-            this.matcher = matcher;
-        }
-
-        public String getNext() {
-            return next;
-        }
-
-        public void setNext(String next) {
-            this.next = next;
-        }
-
-        @Override
-        public boolean equals(Object o) {
-            if (this == o) return true;
-            if (o == null || getClass() != o.getClass()) return false;
-
-            MethodWrapper that = (MethodWrapper) o;
-
-            if (!method.equals(that.method)) return false;
-            if (pattern != null ? !pattern.equals(that.pattern) : that.pattern != null) return false;
-            if (matcher != null ? !matcher.equals(that.matcher) : that.matcher != null) return false;
-            return next != null ? next.equals(that.next) : that.next == null;
-
-        }
-
-        @Override
-        public int hashCode() {
-            int result = method.hashCode();
-            result = 31 * result + (pattern != null ? pattern.hashCode() : 0);
-            result = 31 * result + (matcher != null ? matcher.hashCode() : 0);
-            result = 31 * result + (next != null ? next.hashCode() : 0);
-            return result;
         }
     }
 }
